Moved ffill, bfill logic to their own files with testing against proper pandas objects
