+++ zipline/history/history_container.py
@@ -19,13 +19,13 @@ from itertools import groupby, product
 import logbook
 import numpy as np
 import pandas as pd
-import pandas.core.common as com
 from six import itervalues, iteritems, iterkeys
 
 from . history import HistorySpec
 
 from zipline.finance.trading import with_environment
 from zipline.utils.data import RollingPanel, _ensure_index
+from zipline.utils.munge import ffill, bfill
 
 logger = logbook.Logger('History Container')
 
@@ -33,64 +33,6 @@ logger = logbook.Logger('History Container')
 # The closing price is referred to by multiple names,
 # allow both for price rollover logic etc.
 CLOSING_PRICE_FIELDS = frozenset({'price', 'close_price'})
-
-
-def _interpolate(values, method, axis=None):
-    if values.ndim == 1:
-        axis = 0
-    elif values.ndim == 2:
-        axis = 1
-    else:
-        raise Exception("Cannot interpolate array with more than 2 dims")
-
-    values = values.copy()
-    values = interpolate_2d(values, method, axis=axis)
-    return values
-
-
-def interpolate_2d(values, method='pad', axis=0, limit=None, fill_value=None):
-    """
-    Copied from the 0.15.2. This did not exist in 0.12.0.
-
-    Differences:
-        - Don't depend on pad_2d and backfill_2d to return values
-        - Removed dtype kwarg. 0.12.0 did not have this option.
-    """
-    transf = (lambda x: x) if axis == 0 else (lambda x: x.T)
-
-    # reshape a 1 dim if needed
-    ndim = values.ndim
-    if values.ndim == 1:
-        if axis != 0:  # pragma: no cover
-            raise AssertionError("cannot interpolate on a ndim == 1 with "
-                                 "axis != 0")
-        values = values.reshape(tuple((1,) + values.shape))
-
-    if fill_value is None:
-        mask = None
-    else:  # todo create faster fill func without masking
-        mask = com.mask_missing(transf(values), fill_value)
-
-    # Note: pad_2d and backfill_2d work inplace in 0.12.0 and 0.15.2
-    # in 0.15.2 they also return a reference to values
-    if method == 'pad':
-        com.pad_2d(transf(values), limit=limit, mask=mask)
-    else:
-        com.backfill_2d(transf(values), limit=limit, mask=mask)
-
-    # reshape back
-    if ndim == 1:
-        values = values[0]
-
-    return values
-
-
-def ffill(values, axis=None):
-    return _interpolate(values, 'pad', axis=axis)
-
-
-def bfill(values, axis=None):
-    return _interpolate(values, 'bfill', axis=axis)
 
 
 def ffill_buffer_from_prior_values(freq,
